apue figure1.3 ls(1) example
This is an example from Chapter 01, Unix System Overview, explaining Files and Directories and demonstrating the Unix ls(1) command.
Like the ls command, it receives directory information from the main function’s arguments and displays that information to the user. In this example, the following standard libraries are used:
env, fs, io

The env library is used to access the arguments passed to main.
The fs library is used to interact with the file system.
The io library is used to handle errors.

Instead of using APUE’s C-style error-handling utilities like err_quit or err_sys, this example adopts a Rust-friendly style by returning io::Result<()>.
Returning a Result allows flexible handling of errors and makes it easy to use with the ? operator.